# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "This is an example API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/style": {
            "get": {
                "tags": [
                    "Style"
                ],
                "summary": "Get all Branch",
                "operationId": "86a8b3fa75e10a99c81c836ff6ee9d9b",
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/style/getStyle": {
            "get": {
                "tags": [
                    "Style"
                ],
                "summary": "Get  Style",
                "operationId": "3f95ea8a8b788937ea40a7f606239a93",
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/style/store": {
            "post": {
                "tags": [
                    "Style"
                ],
                "description": "",
                "operationId": "78992ff6af8e1973ffb91982a8ac9a68",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "restaurant_id": {
                                                "type": "integer"
                                            },
                                            "primary": {
                                                "type": "string"
                                            },
                                            "onPrimary": {
                                                "type": "string"
                                            },
                                            "secondary": {
                                                "type": "string"
                                            },
                                            "onSecondary": {
                                                "type": "string"
                                            },
                                            "enable": {
                                                "type": "string"
                                            },
                                            "disable": {
                                                "type": "string"
                                            },
                                            "background": {
                                                "type": "string"
                                            },
                                            "onBackground": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "restaurant_id": "restaurant_id",
                                    "primary": "primary",
                                    "onPrimary": "onPrimary",
                                    "secondary": "secondary",
                                    "onSecondary": "onSecondary",
                                    "enable": "enable",
                                    "disable": "disable",
                                    "background": "background",
                                    "onBackground": "onBackground"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "restaurant_id": {
                                            "type": "integr",
                                            "example": "restaurant_id"
                                        },
                                        "primary": {
                                            "type": "string",
                                            "example": "primary"
                                        },
                                        "onPrimary": {
                                            "type": "string",
                                            "example": "onPrimary"
                                        },
                                        "secondary": {
                                            "type": "string",
                                            "example": "secondary"
                                        },
                                        "onSecondary": {
                                            "type": "string",
                                            "example": "onSecondary"
                                        },
                                        "enable": {
                                            "type": "string",
                                            "example": "enable"
                                        },
                                        "disable": {
                                            "type": "string",
                                            "example": "disable"
                                        },
                                        "background": {
                                            "type": "string",
                                            "example": "background"
                                        },
                                        "onBackground": {
                                            "type": "string",
                                            "example": "onBackground"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "CredentialsError = 1;Required = 100;MaxLength255 = 101;Unique = 103;"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/style/show/{id}": {
            "get": {
                "tags": [
                    "Style"
                ],
                "summary": "Get  Style",
                "operationId": "07f0f5d7ecd59b5431b51ae3699e6935",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/style/update/{id}": {
            "post": {
                "tags": [
                    "Style"
                ],
                "description": "",
                "operationId": "330f623086e0fef0b27290fb64ae737e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "restaurant_id": {
                                                "type": "integer"
                                            },
                                            "primary": {
                                                "type": "string"
                                            },
                                            "onPrimary": {
                                                "type": "string"
                                            },
                                            "secondary": {
                                                "type": "string"
                                            },
                                            "onSecondary": {
                                                "type": "string"
                                            },
                                            "enable": {
                                                "type": "string"
                                            },
                                            "disable": {
                                                "type": "string"
                                            },
                                            "background": {
                                                "type": "string"
                                            },
                                            "onBackground": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "restaurant_id": "restaurant_id",
                                    "primary": "primary",
                                    "onPrimary": "onPrimary",
                                    "secondary": "secondary",
                                    "onSecondary": "onSecondary",
                                    "enable": "enable",
                                    "disable": "disable",
                                    "background": "background",
                                    "onBackground": "onBackground"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "restaurant_id": {
                                            "type": "integr",
                                            "example": "restaurant_id"
                                        },
                                        "primary": {
                                            "type": "string",
                                            "example": "primary"
                                        },
                                        "onPrimary": {
                                            "type": "string",
                                            "example": "onPrimary"
                                        },
                                        "secondary": {
                                            "type": "string",
                                            "example": "secondary"
                                        },
                                        "onSecondary": {
                                            "type": "string",
                                            "example": "onSecondary"
                                        },
                                        "enable": {
                                            "type": "string",
                                            "example": "enable"
                                        },
                                        "disable": {
                                            "type": "string",
                                            "example": "disable"
                                        },
                                        "background": {
                                            "type": "string",
                                            "example": "background"
                                        },
                                        "onBackground": {
                                            "type": "string",
                                            "example": "onBackground"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "CredentialsError = 1;Required = 100;MaxLength255 = 101;Unique = 103;"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/style/delete/{id}": {
            "delete": {
                "tags": [
                    "Style"
                ],
                "summary": "delete  Style",
                "operationId": "538147287391ce613cbecbb8a62c1cae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}